+++ wagtailvideos/models.py
@@ -1,19 +1,19 @@
 from __future__ import absolute_import, print_function, unicode_literals
 
-import datetime
 import logging
 import mimetypes
 import os
 import os.path
-import re
 import shutil
 import subprocess
 import tempfile
 import threading
+from contextlib import contextmanager
 
 from django.conf import settings
 from django.core.exceptions import SuspiciousFileOperation
 from django.core.files.base import ContentFile
+from django.core.files.temp import NamedTemporaryFile
 from django.core.urlresolvers import reverse
 from django.db import models
 from django.db.models.signals import post_save, pre_delete
@@ -102,15 +102,9 @@ class AbstractVideo(CollectionMember, index.Indexed, models.Model):
         index.FilterField('uploaded_by_user'),
     ]
 
-    def is_stored_locally(self):
-        """
-        Returns True if the image is hosted on the local filesystem
-        """
-        try:
-            self.file.path
-            return True
-        except NotImplementedError:
-            return False
+    def __init__(self, *args, **kwargs):
+        super(AbstractVideo, self).__init__(*args, **kwargs)
+        self._initial_file = self.file
 
     def get_file_size(self):
         if self.file_size is None:
@@ -158,57 +152,6 @@ class AbstractVideo(CollectionMember, index.Indexed, models.Model):
 
     def __str__(self):
         return self.title
-
-    def get_duration(self):
-        if self.duration:
-            return self.duration
-
-        if not ffmpeg_installed():
-            return None
-
-        file_path = self.file.path
-        try:
-            # FIXME prints out extra stuff on travis, pip stderr to dev/null
-            show_format = subprocess.check_output(['ffprobe', file_path, '-show_format', '-v', 'quiet'])
-            show_format = show_format.decode("utf-8")
-            # show_format comes out in key=value pairs seperated by newlines
-            duration = re.findall(r'([duration^=]+)=([^=]+)(?:\n|$)', show_format)[0][1]
-            return datetime.timedelta(seconds=float(duration))
-        except subprocess.CalledProcessError:
-            logger.exception("Getting video duration failed")
-            return None
-
-    def get_thumbnail(self):
-        if self.thumbnail:
-            return self.thumbnail
-
-        if not ffmpeg_installed():
-            return None
-
-        file_path = self.file.path
-        file_name = self.filename(include_ext=False) + '_thumb.jpg'
-
-        try:
-            output_dir = tempfile.mkdtemp()
-            output_file = os.path.join(output_dir, file_name)
-            try:
-                FNULL = open(os.devnull, 'r')
-                subprocess.check_call([
-                    'ffmpeg',
-                    '-v', 'quiet',
-                    '-itsoffset', '-4',
-                    '-i', file_path,
-                    '-vcodec', 'mjpeg',
-                    '-vframes', '1',
-                    '-an', '-f', 'rawvideo',
-                    '-s', '320x240',
-                    output_file,
-                ], stdin=FNULL)
-            except subprocess.CalledProcessError:
-                return None
-            return ContentFile(open(output_file, 'rb').read(), file_name)
-        finally:
-            shutil.rmtree(output_dir, ignore_errors=True)
 
     def save(self, **kwargs):
         super(AbstractVideo, self).save(**kwargs)
@@ -347,6 +290,29 @@ class TranscodingThread(threading.Thread):
             shutil.rmtree(output_dir, ignore_errors=True)
 
 
+@contextmanager
+def get_local_file(file):
+    """
+    Get a local version of the file, downloading it from the remote storage if
+    required. The returned value should be used as a context manager to
+    ensure any temporary files are cleaned up afterwards.
+    """
+    try:
+        with open(file.path):
+            yield file.path
+    except NotImplementedError:
+        _, ext = os.path.splitext(file.name)
+        with NamedTemporaryFile(prefix='wagtailvideo-', suffix=ext) as tmp:
+            try:
+                file.open('rb')
+                for chunk in file.chunks():
+                    tmp.write(chunk)
+            finally:
+                file.close()
+            tmp.flush()
+            yield tmp.name
+
+
 # Delete files when model is deleted
 @receiver(pre_delete, sender=Video)
 def video_delete(sender, instance, **kwargs):
@@ -362,8 +328,17 @@ def video_saved(sender, instance, **kwargs):
 
     if hasattr(instance, '_from_signal'):
         return
-    instance.thumbnail = instance.get_thumbnail()
-    instance.duration = instance.get_duration()
+
+    has_changed = instance._initial_file is not instance.file
+    filled_out = instance.thumbnail is not None and instance.duration is not None
+    if has_changed or not filled_out:
+        with get_local_file(instance.file) as file_path:
+            if has_changed or instance.thumbnail is None:
+                instance.thumbnail = ffmpeg.get_thumbnail(file_path)
+
+            if has_changed or instance.duration is None:
+                instance.duration = ffmpeg.get_duration(file_path)
+
     instance.file_size = instance.file.size
     instance._from_signal = True
     instance.save()